follow up cleanup
